First attempt to get synth working on Windows
This change is the first attempt to get this video synth working on Windows.
(Originally, the code did not cater for translating RGB colour values to the 16
colours apparently available on Windows consoles).

Note - it looks like this is not quite ready still, as for me it produces a
single column of cycling colours in the console. More work will be needed to get
to the bottom of what is causing this.