Added some different screen drawing processes
Added some different drawing processes to see if changing how the screen is accessed speeds up performance on the Raspberry Pi.
It made some very minor differences, but nothing significant.  Using PyPy3 is the best way to speed it up, but it still doesn't give the desired performance on the RaspberryPi.